Refactored the Formatters
