+++ src/cfnlint/core.py
@@ -17,10 +17,9 @@
 import logging
 import sys
 import os
-import json
 import argparse
 import six
-from cfnlint import RulesCollection, Match
+from cfnlint import RulesCollection
 import cfnlint.formatters
 import cfnlint.decode
 import cfnlint.maintenance
@@ -38,7 +37,7 @@ class ArgumentParser(argparse.ArgumentParser):
         self.exit(32, '%s: error: %s\n' % (self.prog, message))
 
 
-def run_cli(filename, template, rules, fmt, regions, override_spec, formatter):
+def run_cli(filename, template, rules, regions, override_spec, formatter):
     """Process args and run"""
 
     if override_spec:
@@ -46,7 +45,7 @@ def run_cli(filename, template, rules, fmt, regions, override_spec, formatter):
 
     matches = run_checks(filename, template, rules, regions)
 
-    print_matches(matches, fmt, formatter)
+    formatter.print_matches(matches)
 
     return get_exit_code(matches)
 
@@ -170,6 +169,8 @@ def get_formatter(fmt):
         elif fmt == 'parseable':
             # pylint: disable=bad-option-value
             formatter = cfnlint.formatters.ParseableFormatter()
+        elif fmt == 'json':
+            formatter = cfnlint.formatters.JsonFormatter()
     else:
         formatter = cfnlint.formatters.Formatter()
 
@@ -212,7 +213,7 @@ def get_template_args_rules(cli_args):
         (template, matches) = cfnlint.decode.decode(filename, args.ignore_bad_template)
 
         if matches:
-            print_matches(matches, fmt, formatter)
+            formatter.print_matches(matches)
             sys.exit(get_exit_code(matches))
 
     # If the template has cfn-lint Metadata but the same options are set on the command-
@@ -246,7 +247,7 @@ def get_template_args_rules(cli_args):
         parser.print_help()
         exit(1)
 
-    return(args, filename, template, rules, fmt, formatter)
+    return(args, filename, template, rules, formatter)
 
 
 def get_default_args(template):
@@ -299,46 +300,3 @@ def run_checks(filename, template, rules, regions):
     matches.sort(key=lambda x: (x.filename, x.linenumber, x.rule.id))
 
     return(matches)
-
-
-def print_matches(matches, fmt, formatter):
-    """Print matches"""
-    if fmt == 'json':
-        print(json.dumps(matches, indent=4, cls=CustomEncoder))
-    else:
-        for match in matches:
-            print(formatter.format(match))
-
-
-class CustomEncoder(json.JSONEncoder):
-    """Custom Encoding for the Match Object"""
-    # pylint: disable=E0202
-    def default(self, o):
-        if isinstance(o, Match):
-            if o.rule.id[0] == 'W':
-                level = 'Warning'
-            else:
-                level = 'Error'
-
-            return {
-                'Rule': {
-                    'Id': o.rule.id,
-                    'Description': o.rule.description,
-                    'ShortDescription': o.rule.shortdesc,
-                    'Source': o.rule.source_url
-                },
-                'Location': {
-                    'Start': {
-                        'ColumnNumber': o.columnnumber,
-                        'LineNumber': o.linenumber,
-                    },
-                    'End': {
-                        'ColumnNumber': o.columnnumberend,
-                        'LineNumber': o.linenumberend,
-                    }
-                },
-                'Level': level,
-                'Message': o.message,
-                'Filename': o.filename,
-            }
-        return {'__{}__'.format(o.__class__.__name__): o.__dict__}
