+++ src/cfnlint/formatters/__init__.py
@@ -14,11 +14,24 @@
   OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
   SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 """
+import json
+from cfnlint import Match
 
+class BaseFormatter(object):
+    """Base Formatter class"""
 
-class Formatter(object):
+    def _format(self, match):
+        """Format the specific match"""
+
+    def print_matches(self, matches):
+        """Output all the matches"""
+        # Output each match on a separate line by default
+        for match in matches:
+            print(self._format(match))
+
+class Formatter(BaseFormatter):
     """Generic Formatter"""
-    def format(self, match):
+    def _format(self, match):
         """Format output"""
         formatstr = u'{0} {1}\n{2}:{3}:{4}\n'
         return formatstr.format(
@@ -29,10 +42,50 @@ class Formatter(object):
             match.columnnumber
         )
 
+class JsonFormatter(BaseFormatter):
+    """Json Formatter"""
 
-class QuietFormatter(object):
+    class CustomEncoder(json.JSONEncoder):
+        """Custom Encoding for the Match Object"""
+        # pylint: disable=E0202
+        def default(self, o):
+            if isinstance(o, Match):
+                if o.rule.id[0] == 'W':
+                    level = 'Warning'
+                else:
+                    level = 'Error'
+
+                return {
+                    'Rule': {
+                        'Id': o.rule.id,
+                        'Description': o.rule.description,
+                        'ShortDescription': o.rule.shortdesc,
+                        'Source': o.rule.source_url
+                    },
+                    'Location': {
+                        'Start': {
+                            'ColumnNumber': o.columnnumber,
+                            'LineNumber': o.linenumber,
+                        },
+                        'End': {
+                            'ColumnNumber': o.columnnumberend,
+                            'LineNumber': o.linenumberend,
+                        }
+                    },
+                    'Level': level,
+                    'Message': o.message,
+                    'Filename': o.filename,
+                }
+            return {'__{}__'.format(o.__class__.__name__): o.__dict__}
+
+    def print_matches(self, matches):
+        # JSON formatter outputs a single JSON object
+        print(json.dumps(matches, indent=4, cls=self.CustomEncoder))
+
+
+class QuietFormatter(BaseFormatter):
     """Quiet Formatter"""
-    def format(self, match):
+    def _format(self, match):
         """Format output"""
         formatstr = u'{0} {1}:{2}'
         return formatstr.format(
@@ -42,9 +95,9 @@ class QuietFormatter(object):
         )
 
 
-class ParseableFormatter(object):
-    """Parseable Formmatter"""
-    def format(self, match):
+class ParseableFormatter(BaseFormatter):
+    """Parseable Formatter"""
+    def _format(self, match):
         """Format output"""
         formatstr = u'{0}:{1}:{2}:{3}:{4}:{5}:{6}'
         return formatstr.format(
